sql/analyzer: split rules in multiple files
