+++ sql/analyzer/resolve_functions.go
@@ -1,0 +1,46 @@
+package analyzer
+
+import (
+	"gopkg.in/src-d/go-mysql-server.v0/sql"
+	"gopkg.in/src-d/go-mysql-server.v0/sql/expression"
+)
+
+func resolveFunctions(ctx *sql.Context, a *Analyzer, n sql.Node) (sql.Node, error) {
+	span, ctx := ctx.Span("resolve_functions")
+	defer span.Finish()
+
+	a.Log("resolve functions, node of type %T", n)
+	return n.TransformUp(func(n sql.Node) (sql.Node, error) {
+		a.Log("transforming node of type: %T", n)
+		if n.Resolved() {
+			return n, nil
+		}
+
+		return n.TransformExpressionsUp(func(e sql.Expression) (sql.Expression, error) {
+			a.Log("transforming expression of type: %T", e)
+			if e.Resolved() {
+				return e, nil
+			}
+
+			uf, ok := e.(*expression.UnresolvedFunction)
+			if !ok {
+				return e, nil
+			}
+
+			n := uf.Name()
+			f, err := a.Catalog.Function(n)
+			if err != nil {
+				return nil, err
+			}
+
+			rf, err := f.Call(uf.Arguments...)
+			if err != nil {
+				return nil, err
+			}
+
+			a.Log("resolved function %q", n)
+
+			return rf, nil
+		})
+	})
+}
